Additional edits related to #70 - other ulong to IntPtr
Changes in other interfaces to R library to use IntPtr instead of ulong.
This addresses issues with 32-bit problems reported in issue #70.

+++ R.NET/CharacterVector.cs
@@ -69,7 +69,7 @@ namespace RDotNet
             // To work with ALTREP (introduced in R 3.5.0) and non-ALTREP objects, we will get strings
             // via STRING_ELT, instead of offseting the DataPointer.  This lets R manage the details of
             // ALTREP conversion for us.
-            IntPtr objPointer = GetFunction<STRING_ELT>()(this.DangerousGetHandle(), (ulong)index);
+            IntPtr objPointer = GetFunction<STRING_ELT>()(this.DangerousGetHandle(), (IntPtr)index);
             if (objPointer == Engine.NaStringPointer)
             {
                 return null;
